db_consumer/requirements.txt
kafka-python
six
psycopg2-binary